trying int64_t

+++ src/kendallc.cpp
@@ -235,9 +235,9 @@ NumericVector ici_kt(NumericVector x, NumericVector y, String perspective = "loc
   //return x4;
   IntegerVector obs = compare_both(x4, y4);
   //return obs;
-  long int sum_obs = sum(obs);
+  int64_t sum_obs = sum(obs);
   IntegerVector cnt = diff(which_notzero(obs));
-  long int dis = kendall_discordant(x4, y4);
+  int64_t dis = kendall_discordant(x4, y4);
   
   long double ntie = sum((cnt * (cnt - 1)) / 2);
   // three values should be read as:
@@ -252,7 +252,7 @@ NumericVector ici_kt(NumericVector x, NumericVector y, String perspective = "loc
   double y0 = y_counts[1];
   double y1 = y_counts[2];
   
-  long long int tot = (n_entry * (n_entry - 1)) / 2;
+  int64_t tot = (n_entry * (n_entry - 1)) / 2;
   
   //Note that tot = con + dis + (xtie - ntie) + (ytie - ntie) + ntie
   //              = con + dis + xtie + ytie - ntie
